# [GLUTEN-8704][CH] try accelerate some spark* function by optimizing tight loops