… cpp-ch/local-engine/Functions/SparkFunctionDivide.h
… cpp-ch/local-engine/Parser/ExpressionParser.cpp
… gluten-ut/spark35/src/test/scala/org/apache/gluten/utils/clickhouse/ClickHouseTestSettings.scala