## What changes were proposed in this pull request?

try accelerate some spark* function through vectorization
- sparkDivide: 
   - vectorVector(1.09x)
   - vectorConstant(25x) 
- castFloatToInt*: 2.67x
- checkDecimalOverflowSparkOrNull:
   - decimal to decimal, scale up: 1.47x
   - decimal to decimal, scale down: 1.94x
   - decimal to decimal, no scale change: 274x
   - int to decimal: 3.07x
   - float to decimal: 10.98x


spark ut "SPARK-28224: Aggregate sum big decimal overflow" was excluded from spark3.5 because it runs in ansi mode, which is not fully supported in gluten currently. 

(Fixes: \#8704)

## How was this patch tested?

existing uts


## performance tests
baseline version
```
---------------------------------------------------------------------------------------
Benchmark                                             Time             CPU   Iterations
---------------------------------------------------------------------------------------
BM_OptSparkDivide_VectorVector                   133184 ns       133174 ns         5328
BM_OptSparkDivide_VectorConstant                  34535 ns        34534 ns        19413
BM_OptSparkDivide_ConstantVector                 106590 ns       106575 ns         6717
BM_OptSparkCastFloatToInt                        185191 ns       185165 ns         3536
BM_OptCheckDecimalOverflowSparkFromDecimal1      737342 ns       737312 ns          951
BM_OptCheckDecimalOverflowSparkFromDecimal2      653192 ns       653165 ns         1073
BM_OptCheckDecimalOverflowSparkFromDecimal3      463498 ns       463486 ns         1511
BM_OptCheckDecimalOverflowSparkFromInt          1025330 ns      1025289 ns          682
BM_OptCheckDecimalOverflowSparkFromFloat        2084486 ns      2084187 ns          330
```

optimized version
```
---------------------------------------------------------------------------------------
Benchmark                                             Time             CPU   Iterations
---------------------------------------------------------------------------------------
BM_OptSparkDivide_VectorVector                   117474 ns       117468 ns         5913
BM_OptSparkDivide_VectorConstant                   1393 ns         1393 ns       503263
BM_OptSparkDivide_ConstantVector                 101636 ns       101634 ns         6890
BM_OptSparkCastFloatToInt                         70839 ns        70836 ns         9840
BM_OptCheckDecimalOverflowSparkFromDecimal1      500702 ns       500677 ns         1000
BM_OptCheckDecimalOverflowSparkFromDecimal2      335018 ns       334988 ns         2091
BM_OptCheckDecimalOverflowSparkFromDecimal3        1690 ns         1690 ns       411554
BM_OptCheckDecimalOverflowSparkFromInt           333732 ns       333702 ns         2092
BM_OptCheckDecimalOverflowSparkFromFloat         189853 ns       189838 ns         3665 
```


speedups
``` txt
BM_OptSparkDivide_VectorVector 133184 117474 1.13373
BM_OptSparkDivide_VectorConstant 34535 1393 24.7918
BM_OptSparkDivide_ConstantVector 106590 101636 1.04874
BM_OptSparkCastFloatToInt 185191 70839 2.61425
BM_OptCheckDecimalOverflowSparkFromDecimal1 737342 500702 1.47262
BM_OptCheckDecimalOverflowSparkFromDecimal2 653192 335018 1.94972
BM_OptCheckDecimalOverflowSparkFromDecimal3 463498 1690 274.259
BM_OptCheckDecimalOverflowSparkFromInt 1025330 333732 3.07232
BM_OptCheckDecimalOverflowSparkFromFloat 2084486 189853 10.9795
```