@taiyang-li 2025-02-13
```
### sparkDivide:  当分母为常数时，加速比24x
优化后
------------------------------------------------------------------------
Benchmark                              Time             CPU   Iterations
------------------------------------------------------------------------
BM_SparkDivide_VectorVector       127921 ns       127916 ns         5409
BM_SparkDivide_VectorConstant       1428 ns         1428 ns       497549
BM_SparkDivide_ConstantVector     109748 ns       109744 ns         6521

优化前
------------------------------------------------------------------------
Benchmark                              Time             CPU   Iterations
------------------------------------------------------------------------
BM_SparkDivide_VectorVector       129609 ns       129605 ns         5408
BM_SparkDivide_VectorConstant      33923 ns        33921 ns        20599
BM_SparkDivide_ConstantVector     105231 ns       105228 ns         6682    


### sparkCastFloattoInt*: float转int32加速比1.67x
优化后
--------------------------------------------------------------------
Benchmark                          Time             CPU   Iterations
--------------------------------------------------------------------
BM_OptSparkCastFloatToInt     112905 ns       112900 ns         6178

优化前
--------------------------------------------------------------------
Benchmark                          Time             CPU   Iterations
--------------------------------------------------------------------
BM_OptSparkCastFloatToInt     188640 ns       188631 ns         3795 



### checkDecimalOverflowSparkOrNull:  decimal转decimal  1.2x
优化后 
---------------------------------------------------------------------------------
Benchmark                                       Time             CPU   Iterations
---------------------------------------------------------------------------------
BM_OptCheckDecimalOverflowSparkOrNull1     619780 ns       619460 ns         1105
BM_OptCheckDecimalOverflowSparkOrNull2     547584 ns       547237 ns         1290

优化前
---------------------------------------------------------------------------------
Benchmark                                       Time             CPU   Iterations
---------------------------------------------------------------------------------
BM_OptCheckDecimalOverflowSparkOrNull1     749359 ns       749337 ns          956
BM_OptCheckDecimalOverflowSparkOrNull2     669309 ns       669282 ns         1075
```