@taiyang-li 2025-02-14
Latest benchmark result:  
```
$ ./build_gcc/utils/extern-local-engine/tests/benchmark_local_engine --benchmark_filter="BM_Opt"      
2025-02-14 12:19:47.636 <Warning> SignalHandler: LD_PRELOAD is not set, SignalHandler is disabled
2025-02-14T12:19:47+08:00
Running ./build_gcc/utils/extern-local-engine/tests/benchmark_local_engine
Run on (32 X 2100 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x16)
  L1 Instruction 32 KiB (x16)
  L2 Unified 1024 KiB (x16)
  L3 Unified 11264 KiB (x2)
Load Average: 4.45, 4.27, 4.08
---------------------------------------------------------------------------------------
Benchmark                                             Time             CPU   Iterations
---------------------------------------------------------------------------------------
BM_OptSparkDivide_VectorVector                   118944 ns       118939 ns         5951
BM_OptSparkDivide_VectorConstant                   1360 ns         1360 ns       513645
BM_OptSparkDivide_ConstantVector                 101752 ns       101748 ns         6903
BM_OptSparkCastFloatToInt                         70640 ns        70637 ns         9913
BM_OptCheckDecimalOverflowSparkOrNull1           622877 ns       622850 ns         1130
BM_OptCheckDecimalOverflowSparkOrNull2           543753 ns       543736 ns         1313
```